Fix 2 remaining fractional share bugs
models.py:
- TradeRecord.qty: int → float (Pydantic was silently truncating
  fractional shares e.g. 2.37 → 2 on every trade save)

iceberg.py:
- _plan_slices: replace floor division (//) with true division (/)
  so fractional quantities like 0.11 shares split correctly
  instead of producing [0.0, 0.11] → filtered to just [0.11]

+++ models.py
@@ -31,7 +31,7 @@ class TradeRecord(BaseModel):
     id: int | None = None
     ticker: str
     action: TradeAction
-    qty: int
+    qty: float
     entry_price: float
     entry_time: str
     sentiment_score: int | None = None
